+++ src/main/java/de/johni0702/minecraft/bobby/BobbyConfig.java
@@ -10,6 +10,8 @@ public class BobbyConfig {
     private boolean enabled = true;
     @Comment("Do not load block entities (e.g. chests) in fake chunks.\nThese need updating every tick which can add up.\n\nEnabled by default because the render distance for block entities is usually smaller than the server-view distance anyway.")
     private boolean noBlockEntities = true;
+    @Comment("Reduces the light levels in fake chunks so you can tell the difference from real ones.")
+    private boolean taintFakeChunks = false;
     @Comment("Delays the unloading of chunks which are outside your view distance.\nSaves you from having to reload all chunks when leaving the area for a short moment (e.g. cut scenes).\nDoes not work across dimensions.")
     private int unloadDelaySecs = 60;
     @Comment("Delete regions from the cache when they have not been loaded for X days.\n\nThe cache for a given world is cleaned up whenever you disconnect from the server.\nEntire worlds are cleaned up when the game starts and you have not visited them for X days.\n\nSet to -1 to disabled.\nSet to 0 to clean up everything after every disconnect.")
@@ -24,6 +26,7 @@ public class BobbyConfig {
     public BobbyConfig(
             boolean enabled,
             boolean noBlockEntities,
+            boolean taintFakeChunks,
             int unloadDelaySecs,
             int deleteUnusedRegionsAfterDays,
             int maxRenderDistance,
@@ -31,6 +34,7 @@ public class BobbyConfig {
     ) {
         this.enabled = enabled;
         this.noBlockEntities = noBlockEntities;
+        this.taintFakeChunks = taintFakeChunks;
         this.unloadDelaySecs = unloadDelaySecs;
         this.deleteUnusedRegionsAfterDays = deleteUnusedRegionsAfterDays;
         this.maxRenderDistance = maxRenderDistance;
@@ -43,6 +47,10 @@ public class BobbyConfig {
 
     public boolean isEnabled() {
         return enabled;
+    }
+
+    public boolean isTaintFakeChunks() {
+        return taintFakeChunks;
     }
 
     public int getUnloadDelaySecs() {